pyproject.toml
[tool.poetry]
name = "BuffyBot"
version = "0.1.0"
description = ""
authors = ["ekwska <ekwiecinska55@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "*"
python-dotenv = "*"
beautifulsoup4 = "*"
requests = "*"
pandas = "*"
lxml = "*"
tqdm = "^4.60.0"
progress = "^1.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
